refactor: move user to modules folder
