+++ src/app/app.ts
@@ -3,7 +3,7 @@ import http, { IncomingMessage, ServerResponse } from "http";
 import { RESOURCE_USERS } from "./constants";
 import { handleError } from "./lib/handle-error";
 import { isResourceValid } from "./lib/is-resource-valid";
-import { UserController } from "./user/user-controller";
+import { UserController } from "./modules/user/user-controller";
 import { NotFoundError } from "./err";
 
 dotenv.config();
