+++ src/app/modules/user/user-repository.ts
@@ -1,5 +1,5 @@
 
-import { NotFoundError } from "../err";
+import { NotFoundError } from "../../err";
 import { User } from "./user-type";
 
 class UserRepository {
